Data

+++ controllers/auth.controller.js
@@ -1,34 +1,22 @@
 const config = require("../config/auth.config");
 const db = require("../models/");
 const User = db.user;
+const Role = db.role;
 var jwt = require("jsonwebtoken");
 var bcrypt = require("bcryptjs");
-const axios = require("axios");
+const axios = require('axios');
 const Favourite = require("../models/favourite.model");
-const Credentials = require("../models/updateCred.model");
+const Credentials = require("../models/updateCred.model")
 const NFT = require("../models/nft.model");
-var multer = require('multer');
-const fs = require('fs');
-//Image Upload
-var storage = multer.diskStorage({
-  destination: (req, file, cb) => {
-      cb(null, 'uploads')
-  },
-  filename: (req, file, cb) => {
-      cb(null, file.fieldname + '-' + Date.now())
-  }
-});
-
-var upload = multer({ storage: storage });
 const headers = {
-  "x-access-token": process.env.COIN_ACCESS_TOKEN,
+  "x-access-token":process.env.COIN_ACCESS_TOKEN ,
 };
 
 exports.signup = (req, res) => {
   const user = new User({
     email: req.body.email,
     phonenumber: req.body.phonenumber,
-    password: bcrypt.hashSync(req.body.password),
+    password: bcrypt.hashSync(req.body.password)
   });
 
   user.save((err) => {
@@ -40,36 +28,40 @@ exports.signup = (req, res) => {
   });
 };
 
-exports.signin = (req, res) => {
+exports.signin =  (req, res) => {
   User.findOne({
-    email: req.body.email,
-  }).exec((err, user) => {
-    if (err) {
-      res.status(500).send({ message: err });
-      return;
-    }
+    email: req.body.email
+  })
+    .exec((err, user) => {
+      if (err) {
+        res.status(500).send({ message: err });
+        return;
+      }
 
-    if (!user) {
-      return res.status(404).send({ message: "User Not found." });
-    }
+      if (!user) {
+        return res.status(404).send({ message: "User Not found." });
+      }
 
-    var passwordIsValid = bcrypt.compareSync(req.body.password, user.password);
+      var passwordIsValid =  bcrypt.compareSync(
+        req.body.password,
+        user.password
+      );
 
-    if (!passwordIsValid) {
-      return res.status(401).send({
-        accessToken: null,
-        message: "Invalid Password!",
+      if (!passwordIsValid) {
+        return res.status(401).send({
+          accessToken: null,
+          message: "Invalid Password!"
+        });
+      }
+      var token = jwt.sign({ id: user.id }, config.secret, {
+        expiresIn: 86400 // 24 hours
       });
-    }
-    var token = jwt.sign({ id: user.id }, config.secret, {
-      expiresIn: 86400, // 24 hours
+      res.status(200).send({
+        id: user._id,
+        email: user.email,
+        accessToken: token
+      });
     });
-    res.status(200).send({
-      id: user._id,
-      email: user.email,
-      accessToken: token,
-    });
-  });
 };
 
 exports.getdata = async function (req, res, next) {
@@ -84,9 +76,9 @@ exports.getdata = async function (req, res, next) {
 };
 
 exports.getcoinData = async function (req, res, next) {
-  var uuid = req.body.uuid;
-  axios
-    .get("https://api.coinranking.com/v2/coin/" + uuid, { headers })
+  var uuid = req.body.uuid
+   axios
+    .get("https://api.coinranking.com/v2/coin/"+uuid, { headers })
     .then(async function (response) {
       await res.send(response.data);
     })
@@ -117,23 +109,16 @@ exports.fav = (req, res) => {
 };
 
 /*View Favourites*/
-exports.getfav = async function (req, res, next) {
-  const getfav = await Favourite.find({
-    isFav: true,
-    userId: req.body.userId,
-  }).sort({ createdAt: -1 });
+exports.getfav = async function(req, res, next) {
+  const getfav = await Favourite.find({isFav:true,userId:req.body.userId}).sort({createdAt: -1});
   res.send(JSON.stringify(getfav));
-};
+}
 
 /*Delete Favourites*/
-exports.delfav = async function (req, res, next) {
-  const getfav = await Favourite.findOneAndDelete({
-    isFav: true,
-    userId: req.body.userId,
-    name: req.body.name,
-  }).sort({ createdAt: -1 });
+exports.delfav = async function(req, res, next) {
+  const getfav = await Favourite.findOneAndDelete({isFav:true,userId:req.body.userId,name:req.body.name}).sort({createdAt: -1});
   res.send(JSON.stringify(getfav));
-};
+}
 
 /*Add NFT */
 exports.addnft = (req, res) => {
@@ -151,42 +136,27 @@ exports.addnft = (req, res) => {
     res.send({ message: "Added to NFT successfully!" });
   });
 };
-/*Retrieve NFT */
-exports.getnft = async function (req, res, next) {
-  const getnft = await NFT.find({ userId: req.body.userId }).sort({
-    createdAt: -1,
-  });
+ /*Retrieve NFT */
+ exports.getnft = async function(req, res, next) {
+  const getnft = await NFT.find({userId:req.body.userId}).sort({createdAt: -1});
   res.send(JSON.stringify(getnft));
-};
+}
 
-/*Delete NFT */
-exports.delnft = async function (req, res, next) {
-  const delnft = await NFT.findOneAndDelete({
-    userId: req.body.userId,
-    name: req.body.name,
-  }).sort({ createdAt: -1 });
+ /*Delete NFT */
+ exports.delnft = async function(req, res, next) {
+  const delnft = await NFT.findOneAndDelete({userId:req.body.userId,name:req.body.name}).sort({createdAt: -1});
   res.send(JSON.stringify(delnft));
-};
+}
 
 /*Update Personal Credentials*/
 exports.updateCred = (req, res) => {
   const updateCred = new Credentials({
     name: req.body.name,
     address: req.body.address,
-    photo: {
-      data: fs.readFileSync(
-        path.join(__dirname + "/uploads/" + req.file.filename)
-      ),
-      contentType: "image/png",
-    },
+    photo: req.body.photo.filename,
     userId: req.body.userId,
   });
-  updateCred.save((err) => {
-    if (err) {
-      console.log(res.body);
-      res.status(500).send({ message: err });
-      return;
-    }
-    res.send({ message: "Added successfully!" });
-  });
+  console.log(photo);
 };
+
+
